feat(evm): add `transfer_eth` action
Also, remove sign_transfer action
