+++ addons/evm/src/commands/actions/mod.rs
@@ -2,6 +2,7 @@ use std::str::FromStr;
 
 use alloy::primitives::Address;
 use check_confirmations::CHECK_CONFIRMATIONS;
+use transfer_eth::TRANSFER_ETH;
 use txtx_addon_kit::types::stores::ValueStore;
 use txtx_addon_kit::types::{
     commands::PreCommandSpecification, diagnostics::Diagnostic, types::Value, ConstructDid, Did,
@@ -13,7 +14,7 @@ pub mod deploy_contract;
 mod deploy_contract_create2;
 pub mod eth_call;
 pub mod sign_transaction;
-pub mod sign_transfer;
+pub mod transfer_eth;
 pub mod verify_contract;
 
 use call_contract::SIGN_EVM_CONTRACT_CALL;
@@ -21,7 +22,6 @@ use deploy_contract::EVM_DEPLOY_CONTRACT;
 use deploy_contract_create2::EVM_DEPLOY_CONTRACT_CREATE2;
 use eth_call::ETH_CALL;
 use sign_transaction::SIGN_TRANSACTION;
-use sign_transfer::SIGN_EVM_TRANSFER;
 use verify_contract::VERIFY_CONTRACT;
 
 use crate::{
@@ -31,14 +31,14 @@ use crate::{
 
 lazy_static! {
     pub static ref ACTIONS: Vec<PreCommandSpecification> = vec![
-        SIGN_EVM_TRANSFER.clone(),
         SIGN_EVM_CONTRACT_CALL.clone(),
         ETH_CALL.clone(),
         EVM_DEPLOY_CONTRACT.clone(),
         VERIFY_CONTRACT.clone(),
         CHECK_CONFIRMATIONS.clone(),
         SIGN_TRANSACTION.clone(),
-        EVM_DEPLOY_CONTRACT_CREATE2.clone()
+        EVM_DEPLOY_CONTRACT_CREATE2.clone(),
+        TRANSFER_ETH.clone(),
     ];
 }
